move all function entrypoints to functions folder
